requirements.txt
numpy==1.26.4; python_version >= "3.9"
opencv-python==4.9.0.80; python_version >= "3.6"
pyyaml==6.0.1; python_version >= "3.6"
uuid==1.30
